+++ CHANGES.rst
@@ -4,4 +4,7 @@ Changelog
 0.1 (unreleased)
 ----------------
 
+- Organization form fields changes
+  [sgeulette]
 - Initial release
+  [affinitic]